board admin/board 분리
